+++ Sandboxie/core/drv/driver.c
@@ -159,6 +159,8 @@ P_NtCreateToken                 ZwCreateToken               = NULL;
 P_NtCreateTokenEx               ZwCreateTokenEx             = NULL;
 #endif
 
+P_PsIsWin32KFilterEnabledForProcess IsWin32KFilterEnabledForProcess = NULL;
+
 
 //---------------------------------------------------------------------------
 // DriverEntry
@@ -725,6 +727,8 @@ void* Driver_FindMissingService(const char* ProcName, int prmcnt)
 
 _FX BOOLEAN Driver_FindMissingServices(void)
 {
+    UNICODE_STRING uni;
+
     //
     // Retrieve some unexported kernel functions which may be useful
     //
@@ -752,7 +756,6 @@ _FX BOOLEAN Driver_FindMissingServices(void)
     
     /*DbgPrint("Test 1\n");
 
-    UNICODE_STRING uni;
     OBJECT_ATTRIBUTES objattrs;
     RtlInitUnicodeString(&uni, L"\\??\\C:\\Temp\\test.txt");
     InitializeObjectAttributes(&objattrs,
@@ -783,7 +786,7 @@ _FX BOOLEAN Driver_FindMissingServices(void)
 #endif
 
 #ifdef OLD_DDK
-    UNICODE_STRING uni;
+
 	RtlInitUnicodeString(&uni, L"ZwSetInformationToken");
 
     //
@@ -806,6 +809,10 @@ _FX BOOLEAN Driver_FindMissingServices(void)
 		return FALSE;
 	}
 #endif
+
+    RtlInitUnicodeString(&uni, L"PsIsWin32KFilterEnabledForProcess");
+
+    IsWin32KFilterEnabledForProcess = (P_PsIsWin32KFilterEnabledForProcess) MmGetSystemRoutineAddress(&uni);
 
     return TRUE;
 }